Code Files/Multi_Model_D/backend/requirements.txt
django==4.2.11
djangorestframework==3.15.1
pymongo==4.6.3
langchain==0.3.25
langchain-community==0.3.24
langchain-groq==0.2.0
faiss-cpu==1.8.0
pymupdf==1.24.2
opencv-python==4.9.0.80
pillow==10.3.0
pytesseract==0.3.10
django-cors-headers==4.3.1
unstructured==0.12.5